Move spe to separate module
